Velocidad y giro arreglados

+++ JuegoLibreTanques/Player.h
@@ -7,7 +7,7 @@ class Player : public Actor
 public:
 	Player(float x, float y, float angle, Game* game);
 	void update();
-	void move(float axis);
+	void move(int axis);
 	void rotate(float angle);
 };
